[Conformance checking] Put associated type inference into its own file.
Factor associated type inference out into its own separate source file, so
it's easier to maintain on its own.
